front - account - re-organize component files
