+++ front/src/account/fields/AccountFieldEmail.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import gql from "graphql-tag";
-import AccountField from "../AccountField";
+import AccountField from "./AccountField";
 import AccountSimpleFieldForm from "./forms/AccountSimpleFieldForm";
 
 type Me = {
